requirements.txt
faiss-cpu 
langchain 
langchain-google-genai 
google-generativeai 
pypdf 
python-dotenv 
langchain-groq
